+++ nextstrain/cli/runner/conda.py
@@ -1,6 +1,39 @@
 """
 Run commands with access to a fully-managed Conda environment.
 
+`Conda <https://docs.conda.io/en/latest/miniconda.html>`__ is a very popular
+packaging system freely-available for all platforms. When you use Nextstrain
+CLI's built-in Conda support, you don't need to install any other Nextstrain
+software dependencies yourself as validated versions are already bundled into a
+package (`nextstrain-base`_) by the Nextstrain team and automatically managed
+in an isolated location (isolated even from other Conda environments you may
+manage yourself).
+
+.. _nextstrain-base: https://github.com/nextstrain/conda-base
+
+
+.. versionadded:: 5.0.0
+
+
+.. _conda-setup:
+
+Setup
+=====
+
+On macOS and Linux, run ``nextstrain setup conda`` to get started.
+
+This will download compressed packages totaling about 450 MB in size which
+expand to a final on-disk size of about 2 GB.  Transient disk usage during this
+process peaks at about 3 GB.  These numbers are current as of August 2023, as
+observed on Linux.  Numbers will vary over time, with a tendency to slowly
+increase, and vary slightly by OS.
+
+This runtime is not directly supported on Windows, but you can use `WSL2
+<https://docs.microsoft.com/en-us/windows/wsl/wsl2-index>`__ to "switch" to
+Linux and run the above setup command.
+
+
+.. _conda-env:
 
 Environment variables
 =====================